fix rouletteWheel
The roulette wheel's limit is now an attribute of the class, as the roulette wheel. 

This solves an issue in which the limit value was lost after chosing the parents for the first dino and for all the other dinos of the generation, the allocation value was wrongly used.

Maybe it's time to create a class 'RouletteWheel'.