feat: TopNElitism, getTopN
A class for the Elitism Strategy was created. This class applies elitism for the first N dinossaurs in the generation.

In support, a utils function 'getTopN' was created which returns the best N dinossaurs of the generation.

The function 'updateDNADaVezByDinoId' was also moved to utils.

Since we can adapt the number of dinossaurs to apply elitism, a new parameter for this functionality was added to the GenAlgorithm constructor.